refactor(barChart): rename
