+++ src/base/static/components/molecules/bar-chart.js
